Moved 2 more classes to blackbox
