+++ orcid-integration-test/src/test/java/org/orcid/integration/blackbox/api/Orcid2StepOauthFlowTest.java
@@ -14,71 +14,36 @@
  *
  * =============================================================================
  */
-package org.orcid.integration.api.test;
+package org.orcid.integration.blackbox.api;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
-import java.util.Arrays;
-import java.util.List;
-
 import javax.annotation.Resource;
 import javax.ws.rs.core.MultivaluedMap;
 
 import org.codehaus.jettison.json.JSONException;
 import org.codehaus.jettison.json.JSONObject;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.orcid.core.manager.ClientDetailsManager;
 import org.orcid.integration.api.t2.T2OAuthAPIService;
-import org.orcid.test.DBUnitTest;
+import org.orcid.integration.blackbox.BlackBoxBase;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import com.sun.jersey.api.client.ClientResponse;
 import com.sun.jersey.core.util.MultivaluedMapImpl;
 
-/**
- * 
- * @author Angel Montenegro
- *
- */
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(locations = { "classpath:test-oauth-orcid-api-client-context.xml" })
-public class Orcid2StepOauthFlowTest extends DBUnitTest {
+public class Orcid2StepOauthFlowTest extends BlackBoxBase {
 
-    private static final String CLIENT_DETAILS_ID = "APP-5555555555555555";
-    
-    @Resource(name = "t2OAuthClient")
+	@Resource(name = "t2OAuthClient")
     private T2OAuthAPIService<ClientResponse> oauthT2Client;
-    @Resource
-    private ClientDetailsManager clientDetailsManager;
-    
-    private static final List<String> DATA_FILES = Arrays.asList("/data/EmptyEntityData.xml", "/data/SecurityQuestionEntityData.xml", "/data/SourceClientDetailsEntityData.xml", "/data/ProfileEntityData.xml",
-            "/data/WorksEntityData.xml", "/data/ClientDetailsEntityData.xml", "/data/Oauth2TokenDetailsData.xml",
-            "/data/WebhookEntityData.xml");
 
-    @BeforeClass
-    public static void initDBUnitData() throws Exception {
-        initDBUnitData(DATA_FILES);        
-    }
-    
-    @Before
-    public void before() {
-        clientDetailsManager.updateLastModified(CLIENT_DETAILS_ID);
-    }
-    
-    @AfterClass
-    public static void after() throws Exception {
-        removeDBUnitData(DATA_FILES);
-    } 
-    
-    @Test
+	@Test
     public void testWebhook() throws InterruptedException, JSONException {        
         ClientResponse tokenResponse = getClientResponse("/webhook");
         assertEquals(200, tokenResponse.getStatus());
@@ -175,14 +140,13 @@ public class Orcid2StepOauthFlowTest extends DBUnitTest {
         assertNotNull(scope);
         assertEquals("/orcid-profile/create", scope);
     }
-    
+
     private ClientResponse getClientResponse(String scope) {
         MultivaluedMap<String, String> params = new MultivaluedMapImpl();
-        params.add("client_id", CLIENT_DETAILS_ID);
-        params.add("client_secret", "client-secret");
+        params.add("client_id", client1ClientId);
+        params.add("client_secret", client1ClientSecret);
         params.add("grant_type", "client_credentials");
         params.add("scope", scope);
         return oauthT2Client.obtainOauth2TokenPost("client_credentials", params);
     }
-    
 }
